requirements.txt
requests==2.28.1
tweepy==4.10.1
Pillow~=9.2.0
python-dotenv~=0.21.0